requirements.txt
higher==0.2.1
jax==0.4.25
matplotlib==3.6.3
numpy==1.23.5
opacus==1.4.1
opencv_contrib_python==4.7.0.72
opencv_python==4.5.5.64
pandas==1.5.3
Pillow==9.2.0
Pillow==10.2.0
pytorch3d==0.7.0
PyYAML==6.0
PyYAML==6.0.1
scikit_learn==1.2.1
scipy==1.5.2
setuptools==59.6.0
setuptools==59.5.0
six==1.16.0
torch==1.10.0
torchcsprng==0.2.1
torchvision==0.11.2
tqdm==4.64.1
trimesh==3.18.3
tvm==1.0.0
